+++ src/test/converter2/issues/gh1651.ts
@@ -1,0 +1,6 @@
+/** bar */
+export const bar = () => {};
+/** metadata */
+bar.metadata = "baz";
+/** fn */
+bar.fn = () => {};